src/app/requirements.txt
protobuf
grpcio
futures
requests
socket